refactor: move mixins into separate modules
