+++ smarter/smarter/common/mixins/__init__.py
@@ -1,0 +1,11 @@
+"""Common classes"""
+
+from .helper_mixin import SmarterHelperMixin
+from .middleware_mixin import SmarterMiddlewareMixin
+from .singleton import Singleton
+
+__all__ = [
+    "Singleton",
+    "SmarterHelperMixin",
+    "SmarterMiddlewareMixin",
+]
