+++ smarter/smarter/common/mixins/singleton.py
@@ -1,0 +1,24 @@
+"""
+Singleton: A metaclass for creating singleton classes.
+"""
+
+
+class Singleton(type):
+    """
+    A metaclass for creating singleton classes.
+
+    usage:
+    class MyClass(metaclass=Singleton):
+    """
+
+    _instances = {}
+
+    def __call__(cls, *args, **kwargs):
+        if cls not in cls._instances:
+            cls._instances[cls] = super().__call__(*args, **kwargs)
+        return cls._instances[cls]
+
+
+__all__ = [
+    "Singleton",
+]
